restructuring tests
