+++ test/pure/args.fate
