+++ test/pure/assign.fate
