+++ test/pure/basics.fate
