+++ test/pure/calls.fate
