+++ test/pure/do.fate
