+++ test/pure/function.fate
