+++ test/pure/json.fate
