+++ test/pure/lambda.fate
