+++ test/pure/mutable.fate
