+++ test/pure/patterns.fate
