+++ test/pure/reduce.fate
