+++ test/pure/scope.fate
